+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [vX.XX.X] - 2024-XX-XX
 
+### Added
+
+- PT: `at-hash` stamp to store the AT's hash of an invoice
+
 ## [v0.79.0] - 2024-06-05
 
 ### Added