+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
-## [vX.XX.X] - 2024-XX-XX
+## [vX.XX.X] - XXXX-XX-XX
+
+### Added
+
+- `num.AmountFromFloat64` method helps avoid rounding issues when using floats.
+
+### Changed
+
+- MX: Fuel account balance totals calculated maintaining item price precision.
 
 ### Added